pyproject.toml
[project]
name = "turbo-az-mcp"
version = "1.0.0"
description = "Turbo.az MCP Server - Avtomobil axtarışı üçün Claude Desktop inteqrasiyası"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "aiohttp>=3.9.0",
    "Pillow>=10.0.0",
]

[project.scripts]
turbo-az-mcp = "src.server:main"
turbo-az-mcp-http = "src.server_http:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
